requirements.txt
Flask
Flask-Session
requests
psycopg2
Gunicorn
pillow
rembg